+++ R/check-data3.R
@@ -28,8 +28,6 @@ check_data3 <- function(data, values = NULL, min_row = 1, max_row = max_nrow(),
   if (!is.character(data_name)) data_name <- deparse(data_name)
   check_flag_internal(select)
 
-  if (select) warning("argument select is deprecated")
-
   data <- check_data_frame(data, data_name = data_name)
   data <- check_rows(data, min_row = min_row, max_row = max_row, data_name = data_name)
   data <- check_values(data, values = values, unique = TRUE, nulls = FALSE, data_name = data_name)